Created RenderItems folder & added RenderMain component also created PopularData file & updated both Main & Popular components
