+++ Components/Data/PopularData.js
@@ -1,0 +1,40 @@
+const Data = [
+  {
+    name: 'Don Bosco',
+    rating: 5,
+    comment: 'Good',
+    image: require('../assets/avatars/avatar1.png'),
+  },
+  {
+    name: 'Stephen Fleming',
+    rating: 4,
+    comment: 'Ok',
+    image: require('../assets/avatars/avatar2.png'),
+  },
+  {
+    name: 'Julia Roberts',
+    rating: 2,
+    comment: 'Bad',
+    image: require('../assets/avatars/avatar3.png'),
+  },
+  {
+    name: 'Mathew Hayden',
+    rating: 5,
+    comment: 'good food good mood',
+    image: require('../assets/avatars/avatar4.png'),
+  },
+  {
+    name: 'Nathen Astle',
+    rating: 4,
+    comment: 'Very good',
+    image: require('../assets/avatars/avatar2.png'),
+  },
+  {
+    name: 'Jondy Rodes',
+    rating: 2,
+    comment: 'normal',
+    image: require('../assets/avatars/avatar4.png'),
+  },
+];
+
+export default Data;
