+++ Components/MainMenu/Popular/Popular.js
@@ -2,47 +2,10 @@ import React from 'react';
 
 import {View, FlatList, StyleSheet, Text, Image} from 'react-native';
 import ratingStar from '../../Ratings/Ratings';
+import Data from '../../Data/PopularData';
+
 import Icon from 'react-native-vector-icons/Ionicons';
 Icon.loadFont();
-
-const Data = [
-  {
-    name: 'Don Bosco',
-    rating: 5,
-    comment: 'Good',
-    image: require('../../assets/avatars/avatar1.png'),
-  },
-  {
-    name: 'Stephen Fleming',
-    rating: 4,
-    comment: 'Ok',
-    image: require('../../assets/avatars/avatar2.png'),
-  },
-  {
-    name: 'Julia Roberts',
-    rating: 2,
-    comment: 'Bad',
-    image: require('../../assets/avatars/avatar3.png'),
-  },
-  {
-    name: 'Mathew Hayden',
-    rating: 5,
-    comment: 'good food good mood',
-    image: require('../../assets/avatars/avatar4.png'),
-  },
-  {
-    name: 'Nathen Astle',
-    rating: 4,
-    comment: 'Very good',
-    image: require('../../assets/avatars/avatar2.png'),
-  },
-  {
-    name: 'Jondy Rodes',
-    rating: 2,
-    comment: 'normal',
-    image: require('../../assets/avatars/avatar4.png'),
-  },
-];
 
 class Popular extends React.Component {
   state = {
